requirements.txt
click==8.1.3
Flask==2.2.3
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
Pillow==9.4.0
tflite-runtime==2.11.0
Werkzeug==2.2.3
zipp==3.15.0
